@@ -12,7 +12,7 @@
         "PDF"
     ],
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=5.4.0",
         "swiss-payment-slip/swiss-payment-slip": "0.11.*"
     },
     "require-dev": {
